+++ src/views/Home.vue
@@ -1,10 +1,11 @@
 <template>
   <div class="home">
-    <h1>DUMB COMPONENT</h1>
+    <h1>COMPONENTS</h1>
+    <h2>DUMB COMPONENT</h2>
     <Dumb />
     <hr />
     <br />
-    <h1>NC-LOADER COMPONENT</h1>
+    <h2>NC-LOADER COMPONENT</h2>
     <div class="test">
       <nc-loader
         :has-text="true"
@@ -16,6 +17,14 @@
         :text="$t('lang.loading.msg')"
       />
     </div>
+    <hr />
+    <br />
+    <h2>NC-LIST COMPONENT</h2>
+    <div class="test">
+      <nc-list :has-search="true" :items="items" @item-selected="selectItem" />
+    </div>
+    <hr />
+    <br />
     <h1>NC_MODAL</h1>
     <nc-modal 
       :opened = opened 
@@ -36,23 +45,10 @@
           <p>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
           </p>
-          <p>
-            "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?"
-          </p>
-          <p>
-            "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?"
-          </p>
-          <p>
-            "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?"
-          </p>
-          <p>
-            "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?"
-          </p>
           <p>12NYS5</p>
         </template>
         <template v-slot:footer>
           <button>Copiar código</button>
-          <p>dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sut aliquid ex ea</p>
         </template>
       </nc-modal>
     <button  @click="opened = true">Show Modal</button>
@@ -61,8 +57,10 @@
 
 
 <script>
+// @ is an alias to /src
 import Dumb from '@/components/Dumb.vue'
 import ncLoader from '@/components/nc-loader.vue'
+import ncList from '@/components/nc-list.vue'
 import ncModal from '@/components/nc-modal.vue'
 
 export default {
@@ -70,10 +68,21 @@ export default {
   components: {
     Dumb,
     ncLoader,
+    ncList,
     ncModal
   },
   data() {
     return {
+      items: [
+        {
+          country: 'Spain',
+          code: '+34'
+        },
+        {
+          country: 'Sri Lanka',
+          code: '+94'
+        }
+      ],
       showCloseIcon: true,
       hideHeader: false,
       hideFooter: false,
@@ -83,6 +92,12 @@ export default {
       width: '588px',
       height: '630px',
       noVerticallyAligned: false
+    }
+  },
+  methods: {
+    selectItem(e) {
+      // eslint-disable-next-line no-console
+      console.log(e)
     }
   }
 }